refactor(SearchBar): 데이터&상호작용 컴포넌트로 분리, 검색 기능 구현
